www-client/firefox: split up neon and thumb
Right now using clang makes it impossible to allow fpu=neon, but also with gcc thumb needs additional rust-std called thumbv7neon-unknown-linux-gnueabihf. There is still a lot of testing needed to check if it might work to bootstrap that rust-std, but if it does it hasn't worked to detect that additional rust-std at configure when tested last time half a year ago. Situation might have improved, but now known yet. Furthermore, there is not much hope for thumb to ever work with musl, this is it's own can of worms.

This commit should allow it to use clang and fpu=neon, and allows it to turn of thumb for people using musl or those who fail to set up the thumbv7neon-unknown-linux-gnueabihf rust-std

cross compiling with gcc-9 works in all variants